Push Notifications to a SOLEUS Bluetooth 4 Band
This is an Android app which accepts any notifications passed on by the
Buzzby app from the notification area of your phone and sends them to a
connected SOLEUS Bluetooth 4 Band.

The SOLEUS band will buzz your wrist and display the first 48 characters
of the ticker text of the notification.

By using the BuzzBand_SOLEUS app and the Buzzby app together you can get
notifications displayed on your SOLEUS band.

The BuzzBand_SOLEUS app and the Buzzby app are available on Google Play
Store.